+++ Minio/Helper/OperationsHelper.cs
@@ -315,34 +315,3 @@ public partial class MinioClient
         return fullErrorsList;
     }
 }
-
-public static class OperationsUtil
-{
-    private static readonly List<string> SupportedHeaders = new()
-    {
-        "cache-control", "content-encoding", "content-type",
-        "x-amz-acl", "content-disposition", "x-minio-extract"
-    };
-
-    private static readonly List<string> SSEHeaders = new()
-    {
-        "X-Amz-Server-Side-Encryption-Customer-Algorithm",
-        "X-Amz-Server-Side-Encryption-Customer-Key",
-        "X-Amz-Server-Side-Encryption-Customer-Key-Md5",
-        Constants.SSEGenericHeader,
-        Constants.SSEKMSKeyId,
-        Constants.SSEKMSContext
-    };
-
-    internal static bool IsSupportedHeader(string hdr, IEqualityComparer<string> comparer = null)
-    {
-        comparer ??= StringComparer.OrdinalIgnoreCase;
-        return SupportedHeaders.Contains(hdr, comparer);
-    }
-
-    internal static bool IsSSEHeader(string hdr, IEqualityComparer<string> comparer = null)
-    {
-        comparer ??= StringComparer.OrdinalIgnoreCase;
-        return SSEHeaders.Contains(hdr, comparer);
-    }
-}
